refactor: consolidate utility functions into utils/utils.ts
