+++ scripts/tools/extract-tools-github.ts
@@ -4,7 +4,7 @@ import dotenv from 'dotenv';
 
 import type { ToolsData } from '@/types/scripts/tools';
 
-import { pause } from '../utils';
+import { pause } from '../../utils/utils';
 import { logger } from '../utils/logger';
 
 dotenv.config();
